+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 next
 ----
 #. Use `django-social-auth` to authenticate against external providers.
+#. Drop the wizard style of registration. This is required for consistent UX when registering via Facebook.
 
 1.1.2
 -----