+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ next
 ----
 #. Use `django-social-auth` to authenticate against external providers.
 #. Drop the wizard style of registration. This is required for consistent UX when registering via Facebook.
+#. Listings no longer include unpublished items that are referenced by the Content or Pinned fields.
 
 1.1.2
 -----